src/web/requirements.txt
Flask==1.1.2
google-api-core==1.26.3
google-auth==1.30.0
Werkzeug==1.0.1
